Rename normalization dynamic kernels and tests
