+++ benchmark/one-level-arch/kernels/solution/normalization/group_norm_grad/group_norm_grad_dynamic.hpp
