+++ benchmark/one-level-arch/kernels/solution/normalization/group_norm_grad_1d/group_norm_grad_1d_dynamic.hpp
