+++ benchmark/one-level-arch/kernels/solution/normalization/rms_norm/rms_norm_dynamic.hpp
@@ -1,5 +1,5 @@
 // =============================================================================
-// rms_norm_pto.hpp — RMSNorm (one-level PTO)
+// rms_norm_dynamic.hpp — RMSNorm (one-level PTO)
 // =============================================================================
 //
 // Shape dims: A (outer / row), R (reduce / col).
