+++ benchmark/one-level-arch/test/solution/normalization/group_norm_grad/compile.all
@@ -2,7 +2,7 @@
 # Dynamic 4PE GroupNormGrad N=2,C=32,G=8,HxW=2024; tiling is calculated from runtime shape.
 set -euo pipefail
 : "${COMPILER_DIR:?Set COMPILER_DIR to the compiler bin directory}"
-make TESTCASE=group_norm_grad COMPILER_DIR="$COMPILER_DIR" DType=__half N_BATCH=2 C_CH=32 G_GRP=8 HxW_SZ=2024 PE_NUM=4 diss
+make TESTCASE=group_norm_grad_dynamic COMPILER_DIR="$COMPILER_DIR" DType=__half N_BATCH=2 C_CH=32 G_GRP=8 HxW_SZ=2024 PE_NUM=4 diss
 
 # Fixed-shape static 4PE counterpart.
 make TESTCASE=group_norm_grad_static COMPILER_DIR="$COMPILER_DIR" DType=__half N_BATCH=2 C_CH=32 G_GRP=8 HxW_SZ=2024 PE_NUM=4 diss
