+++ benchmark/one-level-arch/test/solution/normalization/group_norm_grad_1d/Makefile
@@ -7,10 +7,10 @@ GB_TILE_D ?= 0
 GB_TILE_G ?= 0
 DEFINES += -DGB_TILE_D=$(GB_TILE_D) -DGB_TILE_G=$(GB_TILE_G)
 ifneq ($(PE_NUM),4)
-$(error group_norm_grad_1d supports only the dynamic 4PE testcase)
+$(error group_norm_grad_1d supports only the 4PE testcases)
 endif
 START_FILE = $(ROOT)/test/common/_start_multipe.s
-ifneq ($(filter group_norm_grad_1d group_norm_grad_1d_static,$(TESTCASE)),)
+ifneq ($(filter group_norm_grad_1d_dynamic group_norm_grad_1d_static,$(TESTCASE)),)
 DEFINES += -DDType=$(DType) -DN_BATCH=$(N_BATCH) -DC_CH=$(C_CH)
 DEFINES += -DG_GRP=$(G_GRP) -DPE_NUM=$(PE_NUM)
 TARGET = $(ELF_HEAD)_$(TESTCASE)_DType$(DType)_N$(N_BATCH)_C$(C_CH)_G$(G_GRP)_PE$(PE_NUM).elf
