+++ benchmark/one-level-arch/test/solution/normalization/group_norm_grad_1d/compile.all
@@ -2,7 +2,7 @@
 # Dynamic 4PE GroupNormGrad1D N=256,C=256,G=8,D=32; tiling is calculated from runtime shape.
 set -euo pipefail
 : "${COMPILER_DIR:?Set COMPILER_DIR to the compiler bin directory}"
-make TESTCASE=group_norm_grad_1d COMPILER_DIR="$COMPILER_DIR" DType=__half N_BATCH=256 C_CH=256 G_GRP=8 PE_NUM=4 diss
+make TESTCASE=group_norm_grad_1d_dynamic COMPILER_DIR="$COMPILER_DIR" DType=__half N_BATCH=256 C_CH=256 G_GRP=8 PE_NUM=4 diss
 
 # Fixed-shape static 4PE counterpart.
 make TESTCASE=group_norm_grad_1d_static COMPILER_DIR="$COMPILER_DIR" DType=__half N_BATCH=256 C_CH=256 G_GRP=8 PE_NUM=4 diss
