+++ benchmark/one-level-arch/test/solution/normalization/group_norm_grad_1d/src/group_norm_grad_1d_dynamic.cpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 #include "fileop.h"
-#include "solution/normalization/group_norm_grad_1d/group_norm_grad_1d.hpp"
+#include "solution/normalization/group_norm_grad_1d/group_norm_grad_1d_dynamic.hpp"
 
 #ifndef DType
 #define DType __half
