+++ benchmark/one-level-arch/test/solution/normalization/rms_norm/compile.all
@@ -2,7 +2,7 @@
 # Dynamic 4PE RMSNorm [512,8192]; tiling is calculated from runtime shape.
 set -euo pipefail
 : "${COMPILER_DIR:?Set COMPILER_DIR to the compiler bin directory}"
-make TESTCASE=rms_norm COMPILER_DIR="$COMPILER_DIR" DType=__half G_A=512 G_R=8192 PE_NUM=4 diss
+make TESTCASE=rms_norm_dynamic COMPILER_DIR="$COMPILER_DIR" DType=__half G_A=512 G_R=8192 PE_NUM=4 diss
 
 # Fixed-shape static 4PE counterpart.
 make TESTCASE=rms_norm_static COMPILER_DIR="$COMPILER_DIR" DType=__half G_A=512 G_R=8192 PE_NUM=4 diss
