+++ benchmark/one-level-arch/test/solution/normalization/rms_norm_binary/compile.all
@@ -2,7 +2,7 @@
 # Dynamic 4PE RMSNorm binary [16,16384]; tiling is calculated from runtime shape.
 set -euo pipefail
 : "${COMPILER_DIR:?Set COMPILER_DIR to the compiler bin directory}"
-make TESTCASE=rms_norm_binary COMPILER_DIR="$COMPILER_DIR" DType=__half G_A=16 G_R=16384 PE_NUM=4 diss
+make TESTCASE=rms_norm_binary_dynamic COMPILER_DIR="$COMPILER_DIR" DType=__half G_A=16 G_R=16384 PE_NUM=4 diss
 
 # Fixed-shape static 4PE counterpart.
 make TESTCASE=rms_norm_binary_static COMPILER_DIR="$COMPILER_DIR" DType=__half G_A=16 G_R=16384 PE_NUM=4 diss
