+++ benchmark/one-level-arch/test/solution/normalization/rms_norm_binary/src/gen_rms_norm_binary_data.py
@@ -20,7 +20,7 @@ SCRIPT_DIR = Path(__file__).resolve().parent
 DEFAULT_CMP_DIR = (
     SCRIPT_DIR.parents[5]
     / "compare"
-    / "solution_normalization_rms_norm_binary_rms_norm_binary_DType__half_gA16_gR16384_PE4"
+    / "solution_normalization_rms_norm_binary_rms_norm_binary_dynamic_DType__half_gA16_gR16384_PE4"
 )
 DATA_DIR = SCRIPT_DIR / "data"
